# imaxx-smart-office-be-dev/docs/swagger.yaml
basePath: /api
definitions:
...
  request.CreateRoleRequest:
    properties:
      child_role_ids:
        items:
          type: string
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
      permission_ids:
        items:
          type: string
        type: array
      status:
        type: integer
    required:
    - child_role_ids
    - name
    - permission_ids
    - status
    type: object
...
  request.UpdateRoleRequest:
    properties:
      child_role_ids:
        items:
          type: string
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
      permission_ids:
        items:
          type: string
        type: array
      status:
        type: integer
    required:
    - child_role_ids
    - name
    - permission_ids
    - status
    type: object
...
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: iMaxx Office API with Swagger
  title: iMaxx Office API
  version: "1.0"
paths:
...
  /notifications:
    get:
      consumes:
      - application/json
      description: find all unread user notification
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindAll notification
      tags:
      - notifications
  /notifications/set-read/{id}:
    get:
      consumes:
      - application/json
      description: Set read status of notification by id
      parameters:
      - description: Notification ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: SetReadById notification
      tags:
      - notifications
  /notifications/set-read/all:
    patch:
      consumes:
      - application/json
      description: Set read status of notification to all noti
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Set Read All notification
      tags:
      - notifications
  /notifications/unread:
    get:
      consumes:
      - application/json
      description: find all unread user notification
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindUnreadAll notification
      tags:
      - notifications
...
  /roles:
    get:
      consumes:
      - application/json
      description: find all role
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindAllRole role
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: create role
      parameters:
      - description: create role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: CreateRole role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: delete role
      parameters:
      - description: Role ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: DeleteRole role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: find by id role
      parameters:
      - description: Role ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindById role
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: update role
      parameters:
      - description: update role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateRoleRequest'
      - description: Role ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: UpdateRole role
      tags:
      - roles
  /roles/active:
    get:
      consumes:
      - application/json
      description: find all active role
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindActiveAllRole role
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"